AT-195 refactor Exports structure
